+++ version.json
@@ -1,0 +1,5 @@
+{
+    "major": 1,
+    "minor": 2,
+    "patch": 2
+}
